// frontend/projects/vacme-web-generated/src/lib/model/settings-jax.ts
/**
 * Generated VacMe API
 * Generated using custom templates to be found under vacme-web-generated/src/templates.
 *
 * The version of the OpenAPI document: 999.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TranslatedTextJaxTS } from './translated-text-jax';


export interface SettingsJaxTS { 
    distanceImpfungenDesired?: number;
    distanceImpfungenToleranceBefore?: number;
    distanceImpfungenToleranceAfter?: number;
    minutesBetweenInfoUpdates?: number;
    minutesBetweenNumberUpdates?: number;
    showOnboardingWelcomeText?: boolean;
    emailKorrekturEnabled?: boolean;
    generalInfotexts?: Array<TranslatedTextJaxTS>;
    geocodingEnabled?: boolean;
    boosterFreigabeNotificationDisabled?: boolean;
    selbstzahlerFachapplikationEnabled?: boolean;
    selbstzahlerPortalEnabled?: boolean;
    reservationsEnabled: boolean;
    odiFilterWellForGalenicaEnabled: boolean;
    wellUrl: string;
}